+++ example/android/app/src/main/kotlin/com/kjq/fast_mvvm_example/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.kjq.fast_mvvm_example
+package com.kjq.fast_mvvm_example.example
 
 import io.flutter.embedding.android.FlutterActivity