+++ example/lib/main.dart
@@ -1,22 +1,21 @@
-import 'package:flutter/material.dart';
-
 import 'package:fast_mvvm/fast_mvvm.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_easyrefresh/easy_refresh.dart';
 
 import 'article.dart';
 
 void main() {
-  runApp(App());
+  runApp(const App());
 }
 
 class UserModel extends BaseModel {
   Future<bool> login(String account, String psd) async {
-    await Future.delayed(Duration(seconds: 3));
+    await Future.delayed(const Duration(seconds: 3));
     return true;
   }
 
   Future<DataResponse<ArticleEntity>> getArticleList() async {
-    await Future.delayed(Duration(seconds: 1));
+    await Future.delayed(const Duration(seconds: 1));
 
     var entity = ArticleEntity([
       ArticleItem("1", "好的", "内容内容内容内容内容", DateTime.now().toString()),
@@ -30,8 +29,10 @@ class UserModel extends BaseModel {
 }
 
 class App extends StatefulWidget {
+  const App({Key? key}) : super(key: key);
+
   @override
-  _AppState createState() => _AppState();
+  State<StatefulWidget> createState() => _AppState();
 }
 
 class _AppState extends State<App> {
@@ -42,11 +43,11 @@ class _AppState extends State<App> {
       controllerBuild: () => EasyRefreshController(),
       resetRefreshState: (c) =>
           (c as EasyRefreshController).resetRefreshState(),
-      finishRefresh: (c, {bool success = true, bool noMore = false}) =>
+      finishRefresh: (c, {required bool success, required bool noMore}) =>
           (c as EasyRefreshController)
               .finishRefresh(success: success, noMore: noMore),
       resetLoadState: (c) => (c as EasyRefreshController).resetLoadState(),
-      finishLoad: (c, {bool success = true, bool noMore = false}) =>
+      finishLoad: (c, {required bool success, required bool noMore}) =>
           (c as EasyRefreshController)
               .finishLoad(success: success, noMore: noMore),
     );
@@ -55,7 +56,7 @@ class _AppState extends State<App> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       debugShowCheckedModeBanner: false,
       home: SelectPage(),
     );
@@ -68,17 +69,19 @@ class SelectVM extends BaseViewModel {
 }
 
 class SelectPage extends StatelessWidget with BaseView<SelectVM> {
+  const SelectPage({Key? key}) : super(key: key);
+
   @override
   ViewConfig<SelectVM> initConfig() => ViewConfig.noLoad(vm: SelectVM());
 
   @override
-  Widget vmBuild(context, SelectVM vm, Widget? child, Widget? state) {
+  Widget vBuild(context, SelectVM vm, Widget? child, Widget? state) {
     return Scaffold(
-      appBar: AppBar(title: Text("选择")),
+      appBar: AppBar(title: const Text("选择")),
       body: ListView(
         children: <Widget>[
           ListTile(
-            title: Text("是否加载数据,用来测试状态页和重新加载数据"),
+            title: const Text("是否加载数据,用来测试状态页和重新加载数据"),
             trailing: ValueListenableBuilder<bool>(
               valueListenable: vm.isLoadData,
               builder: (_, value, __) => Switch(
@@ -88,7 +91,7 @@ class SelectPage extends StatelessWidget with BaseView<SelectVM> {
             ),
           ),
           ListTile(
-            title: Text("是否单独配置状态页,用来测试状态页和重新加载数据"),
+            title: const Text("是否单独配置状态页,用来测试状态页和重新加载数据"),
             trailing: ValueListenableBuilder<bool>(
               valueListenable: vm.isConfigState,
               builder: (_, value, __) => Switch(
@@ -98,7 +101,7 @@ class SelectPage extends StatelessWidget with BaseView<SelectVM> {
             ),
           ),
           ListTile(
-            title: Text("根布局刷新"),
+            title: const Text("根布局刷新"),
             onTap: () {
               Navigator.push(
                 context,
@@ -113,7 +116,7 @@ class SelectPage extends StatelessWidget with BaseView<SelectVM> {
             },
           ),
           ListTile(
-            title: Text("根布局不刷新"),
+            title: const Text("根布局不刷新"),
             onTap: () {
               Navigator.push(
                 context,