+++ lib/fast_mvvm.dart
@@ -2,4 +2,10 @@ export 'src/base.dart';
 export 'src/common.dart';
 export 'src/widget.dart';
 
+// 基于MVVM架构设计
+// Entity 数据实体类
+// Model 业务接口 数据来源服务器或本地数据库缓存。
+// ViewModel 给View提供数据 调用Model操作数据 刷新View
+// View 视图页面
+
 class FastMvvm {}